requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
ansi2html==1.8.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.1.0
plotly==5.16.1
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
yarl==1.9.2
